% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_administrative.R
\name{add_admin_metadata}
\alias{add_admin_metadata}
\title{Add administrative metadata}
\usage{
add_admin_metadata(
  id,
  .f = "cl_freq",
  admin_format = "id",
  name = NA_character_,
  description = NA_character_
)
}
\arguments{
\item{id}{The code of the administrative metadata.}

\item{.f}{The constructor function of the administrative metadata.}

\item{admin_format}{The return format of the administraive metadata, defaults to
\code{id} (only the id value is returned), or \code{'list'}, \code{'json'} or
\code{'data.frame'}.}

\item{name}{Name of the administrative metadata.}

\item{description}{Description of the administrative metadata.}
}
\value{
An id, a list, a data.frame or a json formatted administrative metadata record.
}
\description{
Add administrative metadata
}
\details{
SDMX Code list for Frequency 2.1 (\code{CL_FREQ}) frequency values
Version 2.1 – 2021-02-13
}
