% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_descriptive.R
\name{add_contributors}
\alias{add_contributors}
\title{Add Contributors}
\usage{
add_contributors(
  givenName,
  familyName,
  contributorType,
  affiliation = NULL,
  orcid = NULL,
  gnd = NULL,
  nameType = "Person",
  contributors = NULL,
  format = "json"
)
}
\arguments{
\item{givenName}{The given name(s) of the creator or contributor.}

\item{familyName}{The family name of the creator or contributor.}

\item{contributorType}{A mandatory type of contribution from the person or organization.
Must be one of :ContactPerson,  DataCollector, DataCurator, DataManager, Distributor, Editor, Funder,
HostingInstitution, Producer, ProjectLeader, ProjectManager, ProjectMember, RegistrationAgency,
RegistrationAuthority, RelatedPerson, Researcher, ResearchGroup, RightsHolder, Supervisor, Sponsor,
WorkPackageLeader, Other.}

\item{affiliation}{Optional affiliation of person.}

\item{orcid}{Optional ORCiD ID of the person.}

\item{gnd}{Optional GND identifier of the person.}

\item{nameType}{Defaults to \code{"Person"}, alternative is \code{"Organizational"}.}

\item{contributors}{Earlier created list of contributors.  Defaults to \code{NULL}.
If the list exists, then the function will append and not re-create the list.}

\item{format}{Defaults to \code{'json'}, in which case a JSON-formatted single character
string is returned.  The alternative is \code{'list'}.}
}
\value{
A list of contributors, in a list or in a JSON formatted character string.
}
\description{
Add one or more contributor.
}
\details{
The institution or person responsible for collecting, managing,
distributing, or otherwise contributing to the development of the resource.
In DataCite, recommended for discovery:
\href{https://support.datacite.org/docs/datacite-metadata-schema-v44-recommended-and-optional-properties#7-contributor}{DataCite Metadata Schema v4.4 Recommended and Optional Properties 7 Contributor}
}
\seealso{
Other persons functions: 
\code{\link{add_creators}()}
}
\concept{persons functions}
