% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_descriptive.R
\name{add_creators}
\alias{add_creators}
\title{Add Creators}
\usage{
add_creators(
  givenName,
  familyName,
  organizationalName = NULL,
  affiliation = NULL,
  orcid = NULL,
  gnd = NULL,
  nameType = "Person",
  creators = NULL,
  format = "json"
)
}
\arguments{
\item{givenName}{The given name(s) of the creator or contributor.}

\item{familyName}{The family name of the creator or contributor.}

\item{organizationalName}{If the \code{"nameType = 'Organizational'"} than
this parameter must be used.}

\item{affiliation}{Optional affiliation of person.}

\item{orcid}{Optional ORCiD ID of the person.}

\item{gnd}{Optional GND identifier of the person.}

\item{nameType}{Defaults to \code{"Person"}, alternative is \code{"Organizational"}.}

\item{creators}{Earlier created list of creators.  Defaults to \code{NULL}.
If the list exists, then the function will append and not re-create the list.}

\item{format}{Defaults to \code{'json'}, in which case a JSON-formatted single character
string is returned.  The alternative is \code{'list'}.}
}
\value{
A list of creators, in a list or in a JSON formatted character string.
}
\description{
Add one or more creator. The main researchers involved in
producing the data, or the authors of the publication, in priority order.
}
\details{
To supply multiple creators, repeat this property.\cr
Allowed values, examples, other constraints: \cr
May be a corporate/institutional or personal name. \cr
Dublin Core:
\href{https://www.dublincore.org/specifications/dublin-core/cross-domain-attribute/}{A Cross-Domain Attribute Set}

DataCite Metadata Schema Mandatory Properties:
\href{https://support.datacite.org/docs/datacite-metadata-schema-v44-mandatory-properties#2-creator}{DataCite Metadata Schema v4.4 Mandatory Properties 2 Creator}
}
\seealso{
Other persons functions: 
\code{\link{add_contributors}()}
}
\concept{persons functions}
