% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_descriptive.R
\name{add_dates}
\alias{add_dates}
\title{Add Dates}
\usage{
add_dates(
  Date = Sys.Date(),
  Updated = Sys.Date(),
  UpdatedAtSource = NULL,
  Collected = NULL,
  EarliestObservation = NULL,
  LatestObservation = NULL,
  StructuralChangeAtSource = NULL,
  Accepted = NULL,
  Available = NULL,
  Created = NULL,
  Copyrighted = NULL,
  Issued = NULL,
  Submitted = NULL,
  Valid = NULL,
  Withdrawn = NULL,
  dates = NULL,
  format = "json"
)
}
\arguments{
\item{Date}{The main date of the resource, unless otherwise stated, should equal
first issue.}

\item{Updated}{The date of the last update to the resource, when the resource is being added to. May be a range.}

\item{UpdatedAtSource}{The date of the last update to the resource at its original source. May be a range.}

\item{Collected}{The date or date range in which the resource content was collected.
To indicate precise or particular timeframes in which research was conducted.}

\item{StructuralChangeAtSource}{The date of the last structural data chanage at source.}

\item{Accepted}{The date that the publisher accepted the resource into their system.
To indicate the start of an embargo period, use Submitted or Accepted, as appropriate.}

\item{Available}{The date the resource is made publicly available. May be a range.
To indicate the end of an embargo period, use Available.}

\item{Created}{The date the resource itself was put together;
this could refer to a timeframe in ancient history, be a date range or a single date for a final component,
e.g., the finalized file with all of the data.
Recommended for discovery.}

\item{Copyrighted}{The specific, documented date at which the resource receives
a copyrighted status, if applicable.}

\item{Issued}{The date that the resource is published or distributed e.g. to a data centre.}

\item{Submitted}{The date the creator submits the resource to the publisher.
This could be different from Accepted if the publisher then applies a selection process.
To indicate the start of an embargo period, use Submitted or Accepted, as appropriate.
Recommended for discovery.}

\item{Valid}{The date or date range during which the dataset or resource is accurate.}

\item{Withdrawn}{The date the resource is removed. It's good practice to indicate the reason for retraction
of withdrawal in the descriptionType.}

\item{format}{format Defaults to \code{'json'}, in which case a JSON-formatted single character
string is returned.  The alternative is \code{'list'} or \code{'character'} (for single Date only.)}

\item{FirstObservation}{The date of the first actual observation (which is not estimated.)}

\item{LastObservation}{The date of the last actual observation (which is not forecasted.)}
}
\description{
Add dates to the administrative and descriptive metadata.
}
\details{
See the controlled date items in
\href{https://support.datacite.org/docs/datacite-metadata-schema-v44-recommended-and-optional-properties#8a-datetype}{DataCite Metaadata Schema 4.4}.
#'
}
\examples{
add_dates()
}
