% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_descriptive.R
\name{add_description}
\alias{add_description}
\title{Add description}
\usage{
add_description(
  Abstract,
  Methods = NULL,
  SeriesInformation = NULL,
  TableOfContents = NULL,
  TechnicalInfo = NULL,
  Other = NULL,
  format = "json"
)
}
\arguments{
\item{Abstract}{A brief description of the resource and the context in which the resource was created.
Recommended for discovery.\cr
Use \code{<br>} to indicate a line break for improved rendering of multiple paragraphs,
but otherwise no html markup.}

\item{Methods}{The methodology employed for the study or research.
Recommended for discovery.}

\item{SeriesInformation}{Information about a repeating series, such as volume,
issue, number.}

\item{TechnicalInfo}{Detailed information that may be associated with design,
implementation, operation, use, and/or maintenance of a process or system.
Defaults to adding \code{\link{add_sessioninfo}}.}

\item{Other}{Other description information that does not fit into an existing category.}

\item{TableofContents}{A listing of the Table of Contents.
Use  \code{<br>} to indicate a line break for improved rendering of multiple paragraphs,
but otherwise no html markup.}
}
\value{

}
\description{
All additional information that does not
fit in any of the other categories. May be used for technical information.
}
\details{
See \href{https://support.datacite.org/docs/schema-optional-properties-v41#17-description}{DataCite Metadata Schema v4.4 Recommended and Optional Properties 14 Description}
}
\examples{
add_description ("My Description")
}
