% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_descriptive.R
\name{add_geolocation}
\alias{add_geolocation}
\title{Add Geolocation}
\usage{
add_geolocation(dat, geo_var = "geo", geo_name_var = NULL, format = "json")
}
\arguments{
\item{dat}{The dataset that needs size information.}

\item{geo_var}{Defaults to \code{geo}.}

\item{format}{Defaults to \code{'json'}, in which case a JSON-formatted single character
string is returned.  The alternative is \code{'list'}.}
}
\value{
A string explaining the file size in CSV format.
}
\description{
Add the spatial region or named place where
the data was gathered or about which the data is focused.
}
\seealso{
Other descriptive_metadata functions: 
\code{\link{add_size}()},
\code{\link{add_titles}()}
}
\concept{descriptive_metadata functions}
