% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_descriptive.R
\name{add_identifiers}
\alias{add_identifiers}
\title{Add identifiders}
\usage{
add_identifiers(
  id = NULL,
  dataset_code = NULL,
  URI = NULL,
  DOI = NULL,
  Version = NULL,
  idAtSource = NULL,
  Other = NULL,
  identifiers = NULL,
  format = "json"
)
}
\arguments{
\item{id}{An identifier string.}

\item{dataset_code}{A short, unique, programatically usable, memorizable dataset
ID, given by an observatory curator.}

\item{URI}{A Uniform Resource Identifier (URI) is a unique sequence of characters that
identifies a logical or physical resource used by web technologies.
URIs may be used to identify anything, including real-world objects,
such as people and places, concepts, or information resources such as web pages and books.}

\item{DOI}{This is the only controlled item in DataCite.}

\item{Version}{A resource version, for example, for an R package.}

\item{idAtSource}{The id used in a source document.}

\item{Other}{Other identifier.}

\item{identifiers}{A list of identifiers that were set earlier. If not the default
\code{NULL} value, the function adds identifiers to this pre-exsisting list of
identifiers.}

\item{format}{Defaults to \code{'json'}, in which case a JSON-formatted single character
string is returned.  The alternative is \code{'list'}.}
}
\value{
A list, a json text of identifier(s).
}
\description{
Add a unique string that identifies a resource.
}
\details{
The Identifier is a unique string that identifies a resource.
For software, determine whether the identifier is for a specific version of a piece of software,
or for all versions

In Dublin Core, Resource Identifier:
\href{https://www.dublincore.org/specifications/dublin-core/cross-domain-attribute/}{A Cross-Domain Attribute Set}

DataCite Metadata Schema Mandatory Properties, Identifer:
\href{https://support.datacite.org/docs/datacite-metadata-schema-v44-mandatory-properties#1-identifier}{DataCite Metadata Schema v4.4 Mandatory Properties 1 Identifier}
}
