% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_administrative.R
\name{add_obs}
\alias{add_obs}
\title{Add observation status}
\usage{
add_obs(id, admin_format = "id")
}
\arguments{
\item{id}{The code of the administrative metadata.}

\item{admin_format}{The return format of the administraive metadata, defaults to
\code{id} (only the id value is returned), or \code{'list'}, \code{'json'} or
\code{'data.frame'}.}
}
\description{
Add observation status coding, such as actual, missing, estimated or
imputed values to a dataset.
}
\details{
Implements SDMX Code List for Observation Status 2.2 (
\href{https://sdmx.org/?sdmx_news=new-version-of-code-list-for-observation-status-version-2-2}{CL_OBS_STATUS)})
frequency values.
Version 2.2, 2021-02-13.
}
