% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wrangling-utils.R
\name{add_observation_status}
\alias{add_observation_status}
\title{Add observation status}
\usage{
add_observation_status(dat)
}
\arguments{
\item{dat}{A data frame.}
}
\description{
Adds the observation status \code{A} for values that are not
missing and \code{O} for missing ones.
}
\details{
The SDMX Code List for Observation Status differentiates \code{O}: Missing value and
\code{M}: Missing value - data cannot exist. Differentiation must be made by the data curator,
as it requires a knowledge of the cause of missingness.
}
\examples{
{
add_observation_status(
  data.frame (
       time = rep(as.Date (paste0(2010:2020, "-01-01")),2),
       geo = c(rep("SE", 11), rep("FI", 11)),
       value = c(1:10, NA_real_, NA_real_, 2:11)
  )
)
}
}
\keyword{export}
