% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_descriptive.R
\name{add_related_items}
\alias{add_related_items}
\title{Add related item}
\usage{
add_related_items(
  RelatedItem,
  relatedItemType = "Dataset",
  relationType = "IsDerivedFrom",
  relatedItemIdentifier,
  relatedItemIdentifierType,
  related_items = NULL,
  format = "json"
)
}
\arguments{
\item{relatedItemType}{\code{Audiovisual}, \code{Book}, \code{BookChapter}, \code{Collection}, \code{ComputationalNotebook}, \code{ConferencePaper}, \code{ConferenceProceeding}, \code{DataPaper}, \code{Dataset}, \code{Dissertation}, \code{Event}, \code{Image}, \code{InteractiveResource}, \code{Journal}, \code{JournalArticle}, \code{Model}, \code{OutputManagementPlan}, \code{PeerReview}, \code{PhysicalObject}, \code{Preprint}, \code{Report}, \code{Service}, \code{Software}, \code{Sound}, \code{Standard}, \code{Text}, \code{Workflow}, \code{Other}}

\item{relationType}{Any of \code{isCitedBy}, \code{Cites}, \code{isSupplementTo}, \code{IsSupplementedBy}, \code{IsDerivedFrom}, \code{isSupplementedBy}, \code{isNewVersionOf}, \code{isPreviousVersionOf}, \code{IsOriginalFormOf}, \code{IsVariantFormOf}, \code{Documents}, \code{IsDocumentedBy}, \code{References}, \code{IsReferencedBy}, \code{IsPublishedIn}, \code{IsPreviousVersionOf}, \code{IsNewVersionOf}, \code{IsVersionOf}, \code{HasVersion}, \code{IsMetadataFor}, \code{HasMetadata}, \code{isPartOf}, \code{IsSourceOf}, \code{hasPart}, \code{compiles}, \code{isCompiledBy}, \code{isIdenticalTo}, \code{isAlternateIdentifier}, \code{Obsoletes}, \code{IsObsoletedBy}, \code{Requires}, \code{IsRequiredBy}, \code{Reviews}, \code{IsReviewedBy}. \cr}

\item{relatedItemIdentifierType}{One of \code{ARK}, \code{arXiv}, \code{bibcode}, \code{DOI}, \code{EAN13}, \code{EISSN}, \code{Handle}, \code{IGSN}, \code{ISBN}, \code{ISSN}, \code{ISTC}, \code{LISSN}, \code{LSID}, \code{PMID}, \code{PURC}, \code{UPC}, \code{URL}, \code{URN}, \code{w3id}.}

\item{relatedItem}{The human readable title or name of the award (grant) as a character string.}

\item{relatedItemIdentifer}{The identifider as a character string.
\code{\link{add_identifiers}}.}
}
\description{
Add observation status coding, such as actual, missing, estimated or
imputed values to a dataset.
}
\details{
In Dublin Core, Relation:
\href{https://www.dublincore.org/specifications/dublin-core/cross-domain-attribute/}{A Cross-Domain Attribute Set}

DataCite Metadata Schema Optional Properties, relatedItem:
\href{https://support.datacite.org/docs/datacite-metadata-schema-v44-recommended-and-optional-properties}{DataCite Metadata Schema v4.4 Recommended and Optional Properties}
}
\examples{
first_rel_item <- add_related_items (
     RelatedItem = "First Related Item",
     relatedItemType = "Dataset",
     relationType = "IsDerivedFrom",
     relatedItemIdentifierType = "Handle",
     relatedItemIdentifier = "first_rel_item",
     related_items = NULL,
  format = 'json' )

add_related_items (
    RelatedItem = "Second Related Item",
    relatedItemType = "Dataset",
    relationType = "IsDerivedFrom",
    relatedItemIdentifierType = "Handle",
    relatedItemIdentifier = "second_rel_item",
    related_items = first_rel_item,
    format = 'json' )
}
