% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_administrative.R
\name{add_sessioninfo}
\alias{add_sessioninfo}
\title{Add session information}
\usage{
add_sessioninfo()
}
\value{
A JSON text.
}
\description{
Converts \code{\link{[utils](sessionInfo)}} into a JSON text.
}
