% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_descriptive.R
\name{add_size}
\alias{add_size}
\title{Add Size}
\usage{
add_size(dat)
}
\arguments{
\item{dat}{The dataset that needs size information.}
}
\value{
A string explaining the file size in CSV format.
}
\description{
Add the file size of the dataset (in CSV format.)
}
\details{
The file will be written to a temporary directory and its sized measured.
It also serves validation purposes for the dataset.
}
\seealso{
Other descriptive_metadata functions: 
\code{\link{add_geolocation}()},
\code{\link{add_titles}()}
}
\concept{descriptive_metadata functions}
