% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_descriptive.R
\name{add_titles}
\alias{add_titles}
\title{Add Titles}
\usage{
add_titles(
  Title = NULL,
  AlternativeTitle = NULL,
  Subtitle = NULL,
  TranslatedTitle = NULL,
  Other = NULL,
  titles = NULL,
  format = "json"
)
}
\arguments{
\item{AlternativeTitle}{Defaults to \code{NULL}, and it is part of the optional
fields of DataCite.}

\item{Subtitle}{Defaults to \code{NULL}.}

\item{TranslatedTitle}{Defaults to \code{NULL}.}

\item{Other}{Defaults to \code{NULL}. Try avoiding its use.}

\item{titles}{A list of earlier set title(s).}

\item{format}{Defaults to \code{'json'}, in which case a JSON-formatted single character
string is returned.  The alternative is \code{'list'}.}

\item{title}{A character string containing the title of the Dataset.
It is mandatory and has no default value. This is part of the Dublin Core and
DataCite mandatory fields.}
}
\value{
A list of titles.
}
\description{
Add a main title, and optional subtitle, alternative- or translated title,
or other title.
}
\details{
Dublin Core:
\href{https://www.dublincore.org/specifications/dublin-core/cross-domain-attribute/}{A Cross-Domain Attribute Set}

DataCite Metadata Schema Mandatory Properties:
\href{https://support.datacite.org/docs/datacite-metadata-schema-v44-mandatory-properties}{DataCite Metadata Schema v4.4 Mandatory Properties}

The \code{Title} field is mandatory in both DataCite and Dublin Core.  DataCite allows multiple
titles to be used, these can be added as a list.
}
\examples{
{
add_titles(title = "My Title",
           AlternativeTitle = "Alternative Title",
           Subtitle = "This is a subtitle",
           TranslatedTitle = "Meine Titel",
           Other = "Other title")
}
}
\seealso{
Other descriptive_metadata functions: 
\code{\link{add_geolocation}()},
\code{\link{add_size}()}
}
\concept{descriptive_metadata functions}
