% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_administrative.R
\name{add_unit}
\alias{add_unit}
\title{Add unit information}
\usage{
add_unit(id, name, admin_format = "id", validate = FALSE)
}
\arguments{
\item{id}{The code of the administrative metadata.}

\item{name}{Name of the administrative metadata.}

\item{admin_format}{The return format of the administraive metadata, defaults to
\code{id} (only the id value is returned), or \code{'list'}, \code{'json'} or
\code{'data.frame'}.}

\item{validate}{Defaults to \code{FALSE}.}
}
\description{
Add observation frequency information to the administrative metadata
of a dataset.
}
\details{
Uses unit codelists.  It has an unvalidated version for free unit data entry.
}
\examples{
add_unit ("NR", "number", admin_format = "json", validate = FALSE)
}
