% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\title{Create a codebook}
\usage{
codebook()
}
\value{
An id, a list, a data.frame or a json formatted administrative metadata record.
}
\description{
Create a codebook
}
\examples{
codebook()
}
