% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook_dataset}
\alias{codebook_dataset}
\alias{is.codebook}
\alias{print.codebook}
\title{Create a codebook for a dataset}
\usage{
codebook_dataset(dataset)

is.codebook(x)

\method{print}{codebook}(x, n = 24, ...)
}
\arguments{
\item{dataset}{A dataset object of class dataset.}

\item{n}{How many rows to print?}
}
\value{
An id, a list, a data.frame or a json formatted administrative metadata record.
}
\description{
Create a codebook for a dataset
}
\examples{
data("small_population")
small_population_dataset <- dataset (
  x= small_population,
  dataset_code = "small_population_total",
  dataset_title = "Population of Small European Countries",
  freq = "A",
  unit = "NR",
  unit_name = "number")

codebook_dataset(small_population_dataset)
}
