% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_discrete_observatory_palette.R
\name{create_discrete_observatory_palette}
\alias{create_discrete_observatory_palette}
\title{Create Observatory Color Palette}
\usage{
create_discrete_observatory_palette(palette_name = "european_countries")
}
\arguments{
\item{palette_name}{Currently only one palette is defined, \code{"european_countries"} which
will create a consistent color palette for European countries.}
}
\value{
A named vector with color codes.
}
\description{
Create consistent palettes for visualizing data.
}
\examples{
{
create_discrete_observatory_palette ("european_countries" )
}
}
\concept{visualisation functions}
