% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_helpers.R
\name{create_json_text}
\alias{create_json_text}
\alias{is.json}
\title{Create JSON String From List}
\usage{
create_json_text(list_to_json)

is.json(text)
}
\arguments{
\item{list_to_json}{A list to be conversted to JSON character string.}

\item{text}{A character string.}
}
\value{
A JSON character string

A logical value, \code{TRUE} for JSON.
}
\description{
Create JSON String From List
}
\examples{
create_json_text (list (id="test_id"))
is.json( '{"id":["test_id"]}')
}
