% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacite.R
\name{datacite}
\alias{datacite}
\alias{datacite_dataset}
\alias{is.datacite}
\alias{print.datacite}
\title{DataCite object}
\usage{
datacite(
  dataset_code,
  Identifier = NA_character_,
  Creator,
  Title,
  Publisher = NA_character_,
  PublicationYear = NA_real_,
  ResourceType = "Dataset",
  Subject,
  Contributor = NA_character_,
  Date = as.character(Sys.Date()),
  Language = "eng",
  AlternateIdentifier = NA_character_,
  RelatedIdentifier = NA_character_,
  Size = NA_character_,
  Format = NA_character_,
  Version = NA_character_,
  Rights = add_rights(),
  Description = NA_character_,
  GeoLocation = NA_character_,
  FundingReference = NA_character_,
  RelatedItem = NA_character_
)

datacite_dataset(
  dataset,
  dataset_code = NULL,
  description = NULL,
  keywords = NULL,
  Title = NULL,
  Subject = NULL,
  Creator,
  Contributor = NULL
)

is.datacite(x)

\method{print}{datacite}(x, ...)
}
\arguments{
\item{dataset_code}{A short, unique, programatically usable, memorizable dataset
ID, given by an observatory curator.}

\item{Identifier}{Identifiers created by \code{\link{add_identifiers}}}

\item{Creator}{Creator field created by \code{\link{add_creators}}.}

\item{Title}{The title of the dataset.}

\item{Publisher}{The publisher of the dataset.}

\item{PublicationYear}{The publication year given as a four-digit number.}

\item{ResourceType}{The type of the resource, defaults to \code{"Dataset"}.}

\item{Subject}{The subject}

\item{Contributor}{The Contributor field created by \code{\link{add_contributors}}.}

\item{Date}{The Date filed created by \code{\link{add_dates}}.}

\item{Language}{Defaults to \code{'eng'}.}

\item{AlternateIdentifier}{Defaults to NA_character_. It should have type subproperty.}

\item{RelatedIdentifier}{Defaults to NA_character_.
Related identifier should have type and relation type sub-properties.}

\item{Size}{The size of the dataset, as measured in CSV format in bytes.}

\item{Format}{The format(s) of the dataset.}

\item{Version}{The version of the dataset.}

\item{Rights}{The rights related to the datset, created by \code{\link{add_rights}}.}

\item{Description}{The description of the dataset.}

\item{dataset}{A dataset object. Its Size, Geolocation and Date properties
will be added to the metadata object.}

\item{description}{A basic description for the Abstract property of the
description.}

\item{keywords}{In the \code{Other} sub-field of the \code{Description} metadata property,
we record keywords for structuring our observatory.}

\item{Geolocation}{The geographical dimension of the observation in the dataset,
created by \code{\link{add_geolocation}}.}

\item{n}{The number of observations to print.}
}
\value{
A data frame of DataCite properties are variable columns.
}
\description{
A metadata object with DataCite mandatory, recommended and optional
fields.
}
\details{
Use \code{\link{datacite_dataset}} to fill up a DataCite record with adding only
those parameters that have no default values. In this case, a given \code{\link{dataset}}
objects:
\itemize{
\item Title will be added as a string.
\item Date will be added with \code{\link{add_dates}} \cr
\item Rigths will be set to the default value by \code{\link{add_rights}} \cr
\item Size will be added by \code{\link{add_size}} \cr
\item GeoLocation will be added by \code{\link{add_geolocation}} \cr
}

Dublin Core:
\href{https://www.dublincore.org/specifications/dublin-core/cross-domain-attribute/}{A Cross-Domain Attribute Set}

Mandatory Properties:
\href{https://support.datacite.org/docs/datacite-metadata-schema-v44-mandatory-properties/}{DataCite Metadata Schema v4.4 Mandatory Properties}

Recommended and Optional Properties:
\href{https://support.datacite.org/docs/datacite-metadata-schema-v44-recommended-and-optional-properties/}{DataCite Metadata Schema v4.4 Recommended and Optional Properties}
}
\examples{
data("small_population")

small_population_dataset <- dataset (
         x= small_population,
         dataset_code = "small_population_total",
         dataset_title = "Population of Small European Countries",
         freq = "A",
         unit = "NR",
         unit_name = "number")

datacite_dataset (
        dataset = small_population_dataset,
        dataset_code = "small_population_total",
        Title = "Population on 1 January by age and sex",
        Subject = "Demography",
        Creator = add_creators("Jane", "Doe", "Reprex")
        )
data("small_population")
small_population_dataset <- dataset (
  x = small_population,
  dataset_code = "small_population_total",
  dataset_title = "Population of Small European Countries",
  freq = "A",
  unit = "NR",
  unit_name = "number")

#The final DataCite properties are Capitalized:

small_population_datacite <- datacite_dataset(
  dataset = small_population_dataset,
  keywords = c("greendeal", "Demography", "Testing"),
  description = "Example dataset with three small countries",
  Subject = "Demography",
  Creator = "Joe, Doe"
  )
}
\seealso{
datacite_properties
}
