% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacite.R
\name{datacite_new}
\alias{datacite_new}
\title{Datacite constructor}
\usage{
datacite_new(
  dataset_code,
  Identifier = NA_character_,
  Creator,
  Title,
  Publisher = NA_character_,
  PublicationYear = NA_real_,
  ResourceType = "Dataset",
  Subject,
  Contributor = NA_character_,
  Date = add_dates(),
  Language = "eng",
  RelatedIdentifier = NA_character_,
  Size = NA_character_,
  Format = NA_character_,
  Version = NA_character_,
  Rights = add_rights(),
  Description = NA_character_,
  GeoLocation = NA_character_,
  FundingReference = NA_character_,
  RelatedItem = NA_character_
)
}
\arguments{
\item{dataset_code}{A short, unique, programatically usable, memorizable dataset
ID, given by an observatory curator.}

\item{Identifier}{Identifiers created by \code{\link{add_identifiers}}}

\item{Creator}{Creator field created by \code{\link{add_creators}}.}

\item{Title}{The title of the dataset.}

\item{Publisher}{The publisher of the dataset.}

\item{PublicationYear}{The publication year given as a four-digit number.}

\item{ResourceType}{The type of the resource, defaults to \code{"Dataset"}.}

\item{Subject}{The subject}

\item{Contributor}{The Contributor field created by \code{\link{add_contributors}}.}

\item{Date}{The Date filed created by \code{\link{add_dates}}.}

\item{Language}{Defaults to \code{'eng'}.}

\item{RelatedIdentifier}{Defaults to NA_character_}

\item{Size}{The size of the dataset, as measured in CSV format in bytes.}

\item{Format}{The format(s) of the dataset.}

\item{Version}{The version of the dataset.}

\item{Rights}{The rights related to the datset, created by \code{\link{add_rights}}.}

\item{Description}{The description of the dataset.}
}
\description{
Datacite constructor
}
\keyword{internal}
