% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacite_properties.R
\docType{data}
\name{datacite_properties}
\alias{datacite_properties}
\title{DataCite Properties}
\format{
A data frame with 20 rows and 3 variables:
\describe{
\item{ID}{ID (a number).}
\item{Property}{Description of the DataCite Field.}
\item{Structure}{Additional strucutre information.}
\item{Obligation}{M - Mandatory, R - Recommended, O - Optional.}
}
}
\source{
\url{https://support.datacite.org/docs/datacite-metadata-schema-v44-properties-overview}
}
\usage{
datacite_properties
}
\description{
The properties of the DataCite Metadata Schema in tabular form.
}
\details{
There are three different levels of obligation for the metadata properties: \cr
\itemize{
\item Mandatory (M) properties must be provided, \cr
\item Recommended (R) properties are optional, but strongly recommended for interoperability \cr
\item Optional (O) properties are optional and provide richer description. \cr
}

More detailed descriptions of the properties, and their related sub-properties,
are provided in:
\itemize{
\item \href{https://support.datacite.org/docs/datacite-metadata-schema-v44-mandatory-properties/}{DataCite Metadata Schema v4.4 Mandatory Properties}; \cr
\item \href{https://support.datacite.org/docs/datacite-metadata-schema-v44-mandatory-properties/}{DataCite Metadata Schema v4.4 Recommended and Optional Metadata}.
}
}
\seealso{
datacite
}
\keyword{datasets}
