% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset}
\alias{dataset}
\alias{is.dataset}
\alias{print.dataset}
\alias{summary.dataset}
\title{dataset for an Automated Data Observatory}
\usage{
dataset(
  x,
  dataset_code,
  dataset_title,
  freq,
  unit,
  unit_name,
  source = "greendeal.dataobservatory.eu"
)

is.dataset(x)

\method{print}{dataset}(x, n = 10, ...)

\method{summary}{dataset}(x, ...)
}
\arguments{
\item{x}{A data frame}

\item{dataset_code}{dataset_code}

\item{dataset_title}{A title}

\item{freq}{A frequency variable coded with \code{\link{add_frequency}}.}

\item{unit}{A unit}

\item{unit_name}{A unit name}

\item{source}{A source}

\item{n}{The number of observations to print.}
}
\description{
The class inherits all methods from a data frame, but has
many fixed attributes.
}
\examples{

my_dataset <- dataset (
     x =data.frame ( time = c(2019,2019,2020,2020),
                geo = c("BE", "BE", "NL", "NL"),
                value = c(10,11,NA,13)),
     dataset_code = "test_code",
     dataset_title = "Test dataset",
     freq = "A",
     unit = "NR", unit_name = "Number"
)
is.dataset(my_dataset)

print(my_dataset,5)
}
