% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate.R
\name{dataset_backcast}
\alias{dataset_backcast}
\title{Backcast the dataset value}
\usage{
dataset_backcast(dataset, backcast_periods = NULL)
}
\arguments{
\item{dataset}{A data frame, tibble or dataset object created by \code{\link{dataset}},
with time, geo, value, frequency, obs_status and method columns.}

\item{forecast_periods}{The number of expected forecasts.  If set to default \code{NULL},
annual datasets will be forecasted to 3 periods, quarterly to 5 periods,
monthly to 12 periods and daily to 30 periods.}
}
\value{
A tibble updated with with forecasted values.
}
\description{
Uses the \code{\link[forecast]{forecast}} function on a dataset.
}
\seealso{
dataset_forecast
}
