% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_reduce.R
\name{dataset_reduce}
\alias{dataset_reduce}
\title{Reduce dataset}
\usage{
dataset_reduce(x, return = "info")
}
\arguments{
\item{x}{A dataset in wide format.}

\item{return}{Any of \code{'data'} for the largest congruent dataset, \code{'info'} for the
properties of the largest congruent dataset, or \code{'both'.}}
}
\value{
A new data frame with the largest congruent dataset, or a info about the congruent dataset,
or a list containing both outputs.
}
\description{
Find the largest congruent sub-dataset of the dataset.
}
\details{
The largest congruent sub-dataset of a dataset is the largest martrix of time and geo
obersvations that do not contain missing values.
}
\examples{
{
tr1 <- data.frame (
    time = c(2010:2020),
    AT = c(NA, 1:10),
    BG = c(NA, NA, 1:6, NA, 8:9),
    CZ = c(NA, 1:9, NA),
    DK = c(rep(NA,10), 1),
    EL = c(rep(NA, 11))
)
dataset_reduce(tr1, return = "both")
}
}
