% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wrangling-utils.R
\name{reduce_dataset}
\alias{reduce_dataset}
\title{Reduce dataset to complete}
\usage{
reduce_dataset(df)
}
\arguments{
\item{df}{A dataset in wide format.}
}
\value{
A new data frame with the largest complete subset.
}
\description{
Find the largest complete subset of the dataset
}
\keyword{internal}
