% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_descriptive.R
\name{related_identifiers_datacite}
\alias{related_identifiers_datacite}
\title{Valid related identifiers}
\usage{
related_identifiers_datacite()
}
\description{
Valid related identifiers
}
\keyword{internal}
