% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-wrangling-utils.R
\name{remove_empty_rows}
\alias{remove_empty_rows}
\title{Remove empty rows}
\usage{
remove_empty_rows(dat)
}
\arguments{
\item{dat}{A data frame, tibble or dataset object created by \code{\link{dataset}},
with time, geo, value, frequency, obs_status and method columns.}
}
\description{
Remove completely empty rows before transforming into ts object.
}
\keyword{internal}
