% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacite.R
\name{validate_datacite}
\alias{validate_datacite}
\title{Validate a datacite object.}
\usage{
validate_datacite(
  dataset_code,
  Identifier,
  Creator,
  Title,
  Publisher,
  PublicationYear,
  ResourceType,
  Subject,
  Contributor,
  Date,
  Language,
  RelatedIdentifier,
  Size,
  Format,
  Version,
  Rights,
  Description,
  GeoLocation,
  FundingReference,
  RelatedItem
)
}
\arguments{
\item{dataset_code}{A short, unique, programatically usable, memorizable dataset
ID, given by an observatory curator.}

\item{Identifier}{Identifiers created by \code{\link{add_identifiers}}}

\item{Creator}{Creator field created by \code{\link{add_creators}}.}

\item{Title}{The title of the dataset.}

\item{Publisher}{The publisher of the dataset.}

\item{PublicationYear}{The publication year given as a four-digit number.}

\item{ResourceType}{The type of the resource, defaults to \code{"Dataset"}.}

\item{Subject}{The subject}

\item{Contributor}{The Contributor field created by \code{\link{add_contributors}}.}

\item{Date}{The Date filed created by \code{\link{add_dates}}.}

\item{Language}{Defaults to \code{'eng'}.}

\item{RelatedIdentifier}{Defaults to NA_character_}

\item{Size}{The size of the dataset, as measured in CSV format in bytes.}

\item{Format}{The format(s) of the dataset.}

\item{Version}{The version of the dataset.}

\item{Rights}{The rights related to the datset, created by \code{\link{add_rights}}.}

\item{Description}{The description of the dataset.}
}
\description{
Validate a datacite object.
}
\keyword{internal}
