% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_descriptive.R
\name{validate_name}
\alias{validate_name}
\title{Validate Names}
\usage{
validate_name(givenName, familyName, organizationalName, nameType)
}
\arguments{
\item{givenName}{The given name(s) of the creator or contributor.}

\item{familyName}{The family name of the creator or contributor.}

\item{organizationalName}{If the \code{"nameType = 'Organizational'"} than
this parameter must be used.}

\item{nameType}{Defaults to \code{"Person"}, alternative is \code{"Organizational"}.}
}
\value{
A validated name as a character string.
}
\description{
Validate Names
}
\keyword{internal}
