% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_freq_value.R
\name{add_freq_value}
\alias{add_freq_value}
\title{Find frequency of the data}
\usage{
add_freq_value(dat)
}
\arguments{
\item{dat}{A data frame with a time variable.}
}
\value{
A tibble with fixed form.
}
\description{
Find frequency of the data
}
\keyword{internal}
