% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_descriptive.R
\name{add_related_items}
\alias{add_related_items}
\title{Add related item}
\usage{
add_related_items(
  RelatedItem,
  relatedItemType = "Dataset",
  relationType = "IsDerivedFrom",
  relatedItemIdentifier,
  related_items = NULL,
  format = "json"
)
}
\arguments{
\item{relatedItemType}{\code{Audiovisual}, \code{Book}, \code{BookChapter}, \code{Collection}, \code{ComputationalNotebook}, \code{ConferencePaper}, \code{ConferenceProceeding}, \code{DataPaper}, \code{Dataset}, \code{Dissertation}, \code{Event}, \code{Image}, \code{InteractiveResource}, \code{Journal}, \code{JournalArticle}, \code{Model}, \code{OutputManagementPlan}, \code{PeerReview}, \code{PhysicalObject}, \code{Preprint}, \code{Report}, \code{Service}, \code{Software}, \code{Sound}, \code{Standard}, \code{Text}, \code{Workflow}, \code{Other}}

\item{relationType}{Any of \code{isCitedBy}, \code{Cites}, \code{isSupplementTo}, \code{IsSupplementedBy}, \code{IsDerivedFrom}, \code{isSupplementedBy}, \code{isNewVersionOf}, \code{isPreviousVersionOf}, \code{IsOriginalFormOf}, \code{IsVariantFormOf}, \code{Documents}, \code{IsDocumentedBy}, \code{References}, \code{IsReferencedBy}, \code{IsPublishedIn}, \code{IsPreviousVersionOf}, \code{IsNewVersionOf}, \code{IsVersionOf}, \code{HasVersion}, \code{IsMetadataFor}, \code{HasMetadata}, \code{isPartOf}, \code{IsSourceOf}, \code{hasPart}, \code{compiles}, \code{isCompiledBy}, \code{isIdenticalTo}, \code{isAlternateIdentifier}, \code{Obsoletes}, \code{IsObsoletedBy}, \code{Requires}, \code{IsRequiredBy}, \code{Reviews}, \code{IsReviewedBy}. \cr}

\item{relatedItem}{The human readable title or name of the award (grant) as a character string.}

\item{relatedItemIdentifer}{An Identifier created by
\code{\link{add_identifiers}}.}
}
\description{
Add observation status coding, such as actual, missing, estimated or
imputed values to a dataset.
}
\details{
In Dublin Core, Relation:
\href{https://www.dublincore.org/specifications/dublin-core/cross-domain-attribute/}{A Cross-Domain Attribute Set}

DataCite Metadata Schema Optional Properties, relatedItem:
\href{https://support.datacite.org/docs/datacite-metadata-schema-v44-recommended-and-optional-properties}{DataCite Metadata Schema v4.4 Recommended and Optional Properties}
}
