% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset}
\alias{dataset}
\alias{is.dataset}
\alias{print.dataset}
\alias{summary.dataset}
\title{dataset for an Automated Data Observatory}
\usage{
dataset(
  x,
  dataset_code,
  dataset_title,
  freq,
  unit,
  unit_name,
  source = "greendeal.dataobservatory.eu",
  doi = NULL,
  add_obs_status = TRUE
)

is.dataset(x)

\method{print}{dataset}(x, n = 10, ...)

\method{summary}{dataset}(x, ...)
}
\arguments{
\item{x}{A data frame}

\item{dataset_code}{A unique dataset id code.}

\item{dataset_title}{A title, which should be the main Title if there are several titles
given by \code{\link{add_titles}}.}

\item{freq}{A frequency variable coded with \code{\link{add_frequency}}.}

\item{unit}{A standardized unit id.}

\item{unit_name}{A unit name.}

\item{source}{A source, currently defaults to \code{"greendeal.dataobservatory.eu"}.}

\item{add_obs_status}{Defaults to \code{TRUE}, when missing \code{obs_status} is added with
\code{\link{add_observation_status}}.}

\item{n}{The number of observations to print.}
}
\value{
A validated dataset object.
}
\description{
The class inherits all methods from a data frame, but has
many fixed attributes.
}
\details{
The attributes of the dataset are
\describe{
\item{dataset_code}{The dataset_code id.}
\item{Title}{The title of the dataaset.}
\item{source}{The data source information.}
\item{unit_name}{The name of the unit variable, which is recorded by its id in the dataset.}
\item{sessionInfo}{The session information added by \code{\link{add_sessioninfo}}.}
}
}
\examples{

my_dataset <- dataset (
     x =data.frame ( time = c(2019,2019,2020,2020),
                geo = c("BE", "BE", "NL", "NL"),
                value = c(10,11,NA,13)),
     dataset_code = "test_code",
     dataset_title = "Test dataset",
     freq = "A",
     unit = "NR", unit_name = "Number"
)
is.dataset(my_dataset)

print(my_dataset,5)
}
