% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_reduce.R
\name{dataset_reduce}
\alias{dataset_reduce}
\title{Reduce dataset}
\usage{
dataset_reduce(x)
}
\arguments{
\item{x}{A dataset in wide format.}
}
\value{
A new data frame with the largest complete subset.
}
\description{
Find the largest complete subset of the dataset.
}
\examples{
{
tr1 <- data.frame (
    time = c(2010:2020),
    AT = c(NA, 1:10),
    BG = c(NA, NA, 1:6, NA, 8:9),
    CZ = c(NA, 1:9, NA),
    DK = c(rep(NA,10), 1),
    EL = c(rep(NA, 11))
)
dataset_reduce(tr1)
}
}
